Move EngineCOmmunicator into util package
